Backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

python-dotenv

pydantic
email-validator

passlib[bcrypt]

python-jose[cryptography]

requests
